WIP: changes suggested in PR
