+++ lib/gql/queries/getWishlistQuery.ts
@@ -4,6 +4,7 @@ const getWishlistQuery = /* GraphQL */ `
   query wishlists($pageSize: Int, $startIndex: Int, $sortBy: String, $filter: String) {
     wishlists(pageSize: $pageSize, startIndex: $startIndex, sortBy: $sortBy, filter: $filter) {
       totalCount
+      pageCount
       items {
         ...wishlist
       }
