+++ lib/helpers/formatDate.ts
@@ -1,0 +1,7 @@
+const formatDate = (date: number) => {
+  const d = new Date(date)
+  const dateString = `${d.getDate()}/${d.getMonth()}/${d.getFullYear()}`
+  return dateString
+}
+
+export default formatDate
